# refine sgd