… paddle/fluid/operators/jit/gen/CMakeLists.txt
… paddle/fluid/operators/jit/more/mkl/CMakeLists.txt
… paddle/fluid/operators/jit/refer/CMakeLists.txt
… paddle/fluid/operators/jit/refer/refer.cc
… python/paddle/fluid/tests/unittests/test_sgd_op.py