pyramiddnn 模型（关闭jmalloc和gprofile） 184s => 165s , 提升 12%

op提升51717 => 35958， 提升43.8%

-------
Before,  51717ms
![image](https://user-images.githubusercontent.com/21351065/53468914-092e1b00-3a97-11e9-9081-da9a7aa6a292.png)

-------
After, 35958ms
![image](https://user-images.githubusercontent.com/21351065/53468949-24008f80-3a97-11e9-9eb4-6c4ee0466293.png)


单独jitkernel看提升3倍左右。
```
 Burning 10 times, Repeat 1000 times.
I0226 10:40:13.300755 22135 benchmark.cc:69] Benchmark kSgd.FP32.CPU
I0226 10:40:18.529037 22135 benchmark.cc:157] Kernel Type kSgd: param_height[750000],param_width[256],grad_height[923],grad_width[256],selected_rows_size[923]: Refer takes 220.904 us; JitCode takes 83.467 us; MKL takes 97.036 us; Target takes 84.01 us;
I0226 10:40:18.957252 22135 benchmark.cc:157] Kernel Type kSgd: param_height[750000],param_width[256],grad_height[801],grad_width[256],selected_rows_size[801]: Refer takes 191.499 us; JitCode takes 72.142 us; MKL takes 83.541 us; Target takes 72.1423 us;
I0226 10:40:19.888864 22135 benchmark.cc:157] Kernel Type kSgd: param_height[750000],param_width[256],grad_height[1724],grad_width[256],selected_rows_size[1724]: Refer takes 414.174 us; JitCode takes 158.513 us; MKL takes 183.174 us; Target takes 156.625 us;

```